app/requirements.txt
Django==2.2.10
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-filter==2.2.0
psycopg2==2.8.5
gunicorn==20.0.4